Unregister events in WhenPcEntersExits
This quest action was registering events in constructor and never unregistering them again.
Events are now registered on initial factory of new action or when restoring save state.
Events are now unregistered when loading a save, starting a new game, or when quest ends.
This action could over time cause a stack overflow exception that combined with another issue would spam log endlessly. Quests that used this action heavily (e.g. Warm Ashes) could accelerate the emergence of this problem just by loading a save a few times in a row.