fix static doors collider size
In addition to using the max of width and depth as door thickness, swap
Y and Z if Y size is less than Z size.

It seems to work fine, but I suspect the logic is not perfectly correct;
I suspect Y and Z actually need to be swapped for grounded doors because
you're actually entering by its top. As if door dimensions were
considered _after_ rotations instead of before, or something like that.
(and that could explain why width and depth can switch roles).