Execute ExtraSpellPts payload on Resume after load
Alternate solution to #1681. Sets correct spell point cap from this effect only during load. It is safe to execute effect payload during load for this effect, whereas it might not be safe to do so globally.
Any other effects impacted by load order like this can also use this solution if safe to do so.