Ensure that the scaling vectors are not initialised with zeros
Tried to use a fsSerializer post-processor but couldn't as the DFBlock is a struct and that means there's no way to adjust the values because the obj is passed by value.
This enables world data override files to continue to work without every single model requiring 3 new float values to be added.