Do not use hit normal to find foe spawning point
Previous algorithm to spawn foes freely would first find a direction
slightly outside of player's field of view, check if there's any
obstacle in that direction, and if so find a spawning point along the
normal of collision point.

Using that normal helps finding a clear location (it should be at least
clear from the obstacle just hit), however it changes the direction of
the spawning point, moving it away from field of view borders; Either
inward (making the foe suddenly appear in front of the player) or
outward (surprize attack, which is not great either).
Also the code checks the distance of the hit point, not spawning point,
making it possible for spawns to happen at a much lower distance than
minDistance constraint.

This patch:
- selects a spawn point along the direction selected at first to be on
the border of player's field of view, so foes should never "appear" in
front of the player (even if sometimes in places that do not make much
logical sense, like before);
- assuming the obstacle is a plane, the new algorithm may choose a spawn
point much closer than before as clear from the obstacle (can't beat
using the normal here);
- however it will always check that spawn happens further than
minDistance, and only add some randomness to the distance if there's
enough room to do so in the first place.