fix rare mobile orientation blind spot
Sometimes, when encountering an idle mobile, it look like it's facing
you until you slightly move around.

The problem is that the choice of mobile orientation is based on cases
using only strict inequalities, so boundary values did not match any
case and orientation was left at its initial value of 0.

I replaced everything with formula, that could be extended more easily
to any number of orientations.