Allow quest to place Item on static quest marker index
This is required for Sx007 "The Werebeast" to assign all quest resources as per classic and historical walkthrough. One letter is placed to a quest marker and the other to an item marker.
Implemented so that default behaviour is observed unless quest specifically calls for "questmarker" index when placing items.