Safer unstick handling using Physics.CapsuleCast()s
Previous unstick handling used a SphereCast to check whether the player
could be teleported forward across some geometry issue.
However the fact that the player controller shape is a capsule instead
of a sphere could lead to an erroneous teleportation inside obstacles
(shop banners, graveyard gate,...)

However just replacing one shape with another is not sufficient, because
hte fact that a controller shaped capsule couldn't move forward is
exactly why the player was stuck in the first place. To be safe, what
should be checked is whether a capsule could freely follow a path to the
place we want to teleport the player to.
Simplest path after the straight line is a step (go up, forward, then
down), that's what this PR does.

Forums: https://forums.dfworkshop.net/viewtopic.php?f=24&t=3291