Support for runtime materials
Added a component that holds a list of archives and records and uses them to assign materials when the prefab is instantiated at runtime. Materials are automatically applied on Awake but can also be manually reapplied.

It allows to "reference" a classic Daggerfall material from a prefab provided by a mod.