Make a cleaner mod manifest file
The mod manifest file is created automatically by Mod Builder, but can also be manually edited to include additional configuration (mod dependencies for example).

When a configuration is not used its value is null. As more options are added, several unnecessary null members could be written to file.

This commit adds a serialization processor that remove null values after serialization.