Add specific keyboard event handlers to UI controls
This is a fix for the issue discussed in #1783:

Keyboard events can leak from game to UI or from UI to game:
- to prevent leaks from UI to game, UI controls can be triggered on
KeyUp, so that the key is released before the game takes control back.
However if the player releases game keys after opening the UI, that can
activate UI buttons ("W" in inventory, "S" in spellbook,...);
- to prevent leaks from game to UI, UI controls can be triggered on
KeyDown, so that keys have to be pressed again inside the UI to have an
effect. However this means keys can still be pressed after the UI
closed, leaking key events into the game.

It should be clear now that those two situations are symetric, and that
none fixes all problems. UI controls need to be able to handle KeyDown
and KeyUp explicitly.

With this PR, buttons can:
- have a OnMouseClick handler; They will only react to mouse events;
- also have a Hotkey hotkey sequence; When the sequence is pressed, the
OnMouseClick handler receives the event as a faked click;
- but now also have a OnKeyboardEvent handler; In that case no click
will be faked, and the keyboard handler will receive both KeyDown and
KeyUp events.

With that in place, controls should react to KeyDown events in general,
but the buttons that close the UI, or maybe even yield control to
another UI window, should use a keyboard handler.
For safety, they should be armed by KeyDown events (possibly giving user
feedback by emitting a sound), and activated by KeyUp events if
previously armed.

I'm very pleased with this, feels like sound design.
I even used it to fix #1805 :)