CullResourceTarget() don't exclude same Place allocation
When using "place npc" or "place item" to move a quest resource, CullResourceTarget() was ignoring cull step if new Place was the same as last Place.
This seems logical on the surface (nothing has changed) but if a quest calls "place npc" twice to the same location then NPC will just get duplicated to that location. This was the case with S0000503.
Remove same-place culling exclusion. This resolves duplicate Baltham in S0000503 on S.14 pathway.