PlayerEnterExit.BuildingDiscoveryData to public set
This property is used to cache building data of current interior for lookups. It is generally set only when entering a building, restoring a save, or teleporting into a building.
But mods might also need to set this at some point, e.g. a custom teleport mod into interiors, and cannot do so when protection set to internal.