Rework movement acceleration setting
The 1-10 scale was much too easy for players to detune movement to the point they were sliding around everywhere and not understanding what was wrong. Even understanding what setting is for, anything below around 9.5 does not feel good.
Reworked this to a simple off/on toggle. Default setting is off and matches classic's zero acceleration stop/go behaviour.
Enabling adds a very small amount of fixed acceleration, equivalent to 9.8 on the old scale. This still allows quick taps of movement keys to make gradual steps, or to jump a smaller distance than is possible without acceleration.