remove emissive? => opaque hack in GhostShader
Now that we have a true transparency and emissive maps to start from,
logic can be moved out of the shader into the logic to create those
maps.
Modded textures will have to be adjusted (on the other hand this gives
modders full freedom over spectral look, which is better in the long
run).