More robust handling of international TEXT.RSC files
Some international TEXT.RSC files (e.g. German) are truncated and do not include full complement of 1408 records.
These RSC files also tend to inject bytecode formatting tokens that are not present in classic. This appears to be an error with how records are encoded by legacy editors which created files.
Handle ID not found, usually due to truncated records.