Enumerate effect templates in first Update()
This was previously done in Start() and localization was not ready in time to lookup string table for effect template names in builds. Moving this to first Update() resolved issue.
Will need to revisit this soon for correctly switching locales at runtime or effect names will remain set to initial language.