Better error message for missing references
Logs the list of missing references if mod loader can't be retrieved from an assembly. For example, this can happen if a mod uses types from another mod that is not installed.

This can be avoided with the dependencies functionality of the mod system, but it's still worth to provide a complete error message.