More accurate pursuit formula
Current pursuit algorithm in EnemySenses compute how much time will be
needed to reach the current position of the target, then computes by how
much the target will have moved by then. This is not very accurate,
because it does both the hypothesis that the target won't move and that
it will move. The fastest the target, the less accurate it gets.

One can exploit that inaccuracy by rushing at enemies, that will turn
their back, pursuing a position behind them. Hence current algorithm has
an additional test to detect that condition, but it doesn't seem to work
perfectly, the exploit still exists.

Accurate approach is to compute at what moment the target will be on a
sphere centered at pursuer position, growing at its interception speed
each second; Or, in 4 dimensions, compute the intersection of a cone and
a line. This is slightly more expensive, solving a quadratic equation
then taking its smallest positive solution, but doesn't seem to suffer
from the previous exploit.