Update viewport when using scale-to-fit large HUD
This changes rendering area to above the HUD only so it is a close match for classic.
When not using scale-to-fit, the large HUD works like a resizable panel on top of fullscreen mode, so viewport does not change.
Considering changing this to "docked" and "undocked" as it better communicates the intent.
TODO: Move crosshair upwards to new viewport centrepoint when large HUD is docked. Check compatibility with retro rendering mode and fix if needed.