Rename large hud settings to docked/undocked
This better communicates the difference when large HUD is docked to bottom of display. This matches classic and updated viewport area.
When large HUD is undocked, it is a floating panel that can be resized and does not change viewport area. This is unique to DFU and obviously doesn't match classic.
Docked large HUD is the default setting.