Move HUD breath bar to own control
This was previously part of HUD vitals and positioned relative to that control's position and scale so would shift to unwanted positions and sizes with flexible large HUD.
This is fixed by making breath bar its own control. Matched position of new breath bar to previous and it remains consistent across all HUD styles.