Adjust baseline jump force down from 8 to 4.5
This is just adequate to jump onto throne lift in Privateer's Hold.
Next change will be for player's Jumping skill to provide additional force up to a set limit. Timing for bunny-hopping will also need to be adjusted.
Baseline jump height now very closely matches the small hops in classic, but will still be higher overall in DFU with a high Jumping skill.