Remove Rigidbody from player
This is an artifact of an older workaround for prior collision bug in Unity. Recent reading suggests that pairing this with a PlayerController can actually cause issues now in Unity 2019.4 LTS.
Removed Rigidbody from player. Should be unnecessary and not cause any additional problems. Tested general collision against spells and environment with no issues.