Set Block values on all child Action Door scripts
This change fixes a bug I found where the Action Doors in a Block don't have their Starting Lock Level (and presumably other variables) set to the block's defined values if the Action Door script is attached to a child of the model's prefab. An example of this is if the door has been replaced by an OBJ model.