Remove text ID 9000 contents
This record is too long for Unity's localized string table editor. Unable to place cursor all the way to end of record and edit text. Also causes odd reformatting of JSON when changing previous records.
This text will be restored soon with each question and answer in an individual record. Requires some refactoring of text importers and class questionnaire to complete.
Note: Class questionnaire will not be functional in master code until this refactor completed. Do not proceed to build until completed.