Added Forward Thrust as an option in Click to Attack
I noticed that the random selection did not include an "UP" option.  I added upRight so that there is a 1 in 6 chance that thrust is selected.  If I added Up instead, it would have been a 3 in 8 chance.

I also modifed FPSWeapon so that Up, UpLeft, and UpRight select the StrikeUp option.  Without this change, UpRight would not have worked.