Update State on all children of Transform Models
This change makes the UpdateMeshRendererDungeonState() function be called on all components of all children of the current transform model instead of only the first component in it's root.

This fixes an issue where already discovered models on the dungeon map are not revealed after loading a save game if an OBJ file has been used to replace the model.