BookFile: Add try-catch check to BinaryReader part
On a few occassions, I've loaded a save only to get the HUD overlaying a black screen. The game still appeared to be running but the game didn't appear to react to my input. I couldn't open the main menu either and had to kill the game. On this occasion, I caught this happening in the editor which reported an EndOfStreamException in the line "header.PageOffsets[i] = reader.ReadUInt32()". As a result, I added a try-catch check to gracefully handle this exception in future for any part of the BookFile's BinaryReader code.