Restore per-task click rearming
Since removing per-task click rearming in 0.11.1 other quests have been found that rely on this behaviour. It appears this was the correct behaviour after all.
There is something more subtle happening in quest Sx011 that I will look at again with fresh eyes. For now, just restoring click rearming behaviour as it was in 0.11.0 and earlier.