Automap: Use right texture table to create models
CreateDungeonGeometryForAutomap() was passing a null value to CreateBaseGameObject() for the texture table. This meant that it used the default texture table to create every dungeon and castle's Automap. This commit changes it so that it will try to get the correct texture table from DaggerfallDungeon before passing a null value.