Added the ability for mods to add new Biography templates. After choosing a class (ex: Mage aka 0), the game looks for all files that match BIOG00T([0-9]+).txt in the BIOG source folder. If multiple files are found, one is selected at random.
In addition, an extension was added to the BIOG file format. Before the first question, if a line starts with #, the rest of the line will be interpreted as a string id, and will be used for the backstory. With a custom ITextProvider, mods can make use of unused string values to provide new backstory text.

For example, if a mod adds a BIOG00T1.txt file, when a player creates a Mage (or similar custom class), the game will randomly pick between BIOG00T0 and BIOG00T1, with possibility for a new set of questions and/or a different backstory.