Bugfix for PR 2053
This fixes a bug in PR 2053 where Runtime Materials only detected that it was being created for a dungeon's or castle's interior if attached to a base object and not if attached to anything else that is inside a dungeon, such as Action Doors. This meant that modded secret doors received the wrong textures. The script will now correctly detect a dungeon/castle interior creation if attached to other objects.