Test console open speed with another font
TEST ONLY.
This change is just to test theory if "Arial" is slowing down console opening, and if using a bundled font speeds up the process.
@petchema Could you please let me know if this resolves the issue for you in #1915. ? If so, we can find a more appropriate font to use here.
Not sure why OpenSans doesn't render correctly, but I'm sure there's any number of free fonts we could use in the console instead if this solves the problem.