Quest and enchanted items exclusions
Sometimes quest items and enchanted items deserve to not be used as
standard items:
- (subreddit below) arrow quest item should not be picked as missile;
- enchanted ingredients should not be used to mix potions (regression in
my refactoring of potion maker);
- I assume quest spell books could exist, and should be given a normal
one as needed;
- I assume quest lanterns could exist, and should be excluded;
- I assume no quest or enchanted wagon/horse, sounds silly;
- I didn't consider quest or enchanted currency either. Currencies are
hard enough (and a bit of a mess already), and also currencies are a
good example of commodity so I don't think quest or enchanted money
would make a great feature.

I also considered using a predicate parameter instead of flags to filter
the result of GetItem() search, it'd be more general but it's also a
matter of taste.

There may be other places where quest and enchanted items could benefit
from exclusions, so I consider this a RFC for the moment.

Subreddit:
https://www.reddit.com/r/daggerfallunity/comments/n07718/til_that_if_your_quest_item_is_an_arrow_using/