Add back Rigidbody to PlayerAdvanced
This was a workaround for older procedural collision issues in Unity between MeshCollider and CharacterController. As people are mentioning this again, worth adding back component for a while to see if this still helps.
This was removed in https://github.com/Interkarma/daggerfall-unity/commit/afa22f919a43f3829c650e500a6ed82feb16d471. Has otherwise been in game for some time.