Added an optional "standard spell index" to effect bundle settings.
When the bundle settings are created from a standard spell, its index is assigned to the settings. When the spell is serialized from the save data, we use this index to update the spellbook's entries to the latest spell record definition, since mods could change it. For spells written to the spellbook before this change, nothing changes on load, and the spell has to be deleted and rewritten by the player to update.