Added new "PrioritizeQuestRumors" experimental setting. When enabled, the TalkManager will show each new quest rumor once before going to the normal rumor mill. Quest rumors are still in the rumor mill, and can be seen again once the "quest rumor priority list" has run out.
As a playthrough advances in DFU, it becomes less and less likely a player would normally see quest rumors due to the sheer volume of region condition rumors. This was accidentally not an issue in classic, but DFU fixed the rumor deletion issue, making the volume of possible rumors much higher.