Destroy extant NPC object when placing it somewhere player is not
In K0C00Y06, the _witness_ NPC is placed at _withouse_ immediately after calling for the guards.

The current implementation of a hot-remove in Place.cs tried to hide the NPC in such an instance but the Tick method of QuestResourceBehavior would un-hide the NPC immediately after. Destroying the NPC's gameobject instead allows for hiding the NPC in a manner that QuestResourceBehavior cannot revert. The NPC will reinstantiate when the player visits its new location.