Increase precision of sound/music slider in launcher
Resolves #2208.
The launcher volume controls had a much lower precision than in game controls. So a low value set in game would be rounded to 0.0 by launcher.
Increased range of sliders in launcher from 0.0 to 100.0 with a reasonable thumb size and more granular resolution.
The two controls still aren't quite identical as they function differently, but the value saved by either control will now be equal to within a decimal place or more.