Exclude guild halls from same-building assignment check
Fixes #2233.
Resolves issue of N0B10Y03 not compiling due to guild questor already being assigned to same location as quest action.
Also fixes other guild quests assigning player to same guildhall as questor.
Another approach is to exclude place resource when it happens to be the building questor resides in. But that's a more complicated check to perform and might undo the intent of same-building check in some cases. As guild halls of a specific faction type are unique within a single location, this is a safer and simpler check to exclude test.
Note that quest can still bounce if guildhall is involved in another *active* quest already running on quest machine. This change only removes issue where questor cannot be in same guild hall as quest action within same quest.