Randomize dye color for clothing items created by quests.
I just had a chat with Jehuty about how this code is used and the glaring problem for clothing item generation is that quest code has no way to check for gender, so it can only ask to generate male or female clothes with no way of knowing if that is correct for the current player character. So I avoid doing any gender checks and came up with this.
From the tests I did, this should produce the correct item each time but with a randomly picked dye.