Scale item based torches after lighting update.
This gives approximately the same illumination as before 0.13.x when the player torch setting is set to 0.7 allowing a bit more configurability for players.

This will need tuning again if we want to adjust the std non-item torchlight intensity.

For info I found that setting dungeon ambient to 0.5 looked about the same as 0.3 before.