Fix for switching back to original enemy Team
The EnemyBasics.Enemies array size is only 62, and the Ids of the objects are not sequential in this array (There is a gap between Lamia Id: 42 and Mage Id: 128).

So just using the Id as the array position would only work for enemies with an Id of 42 or less. Better to check with Linq on the actual Id of the object.