fix seams, take 2
The issue does not seem related to coordinates wrapping, but the fact we
rely on 2 separate computations (texture sampling for tile index on one
side, frac() of uv on the other for tile coordinates) that do not
necessarily match perfectly.

To fix that in a portable and reliable way, use floor()/frac() for both.
Cost is an extra division by _TilemapDim, that's not huge. Maybe
somebody will find a way to get rid of it.