Lower default WeaponAttackThreshold
The current default is too high for players with low mouse DPI and they have troubles initiating swings using gesture attacks.
This can be lowered to 0.001 but that default is much too sensitive for high DPI mouse input.
Does not affect current settings, only new installs and when deploying defaults.