Implementing a UI exit button for controller support
Adding a new button action that allows players to back out of UI screens with a button on the controller, just like how the escape key is used to back out of UI screens on the keyboard.
Key binding can be set in the joystick controller configuration dialog.
Pause key binding renamed in code so it isn't confused with the UI exit binding.