Allow custom TextScale in MessageBox
Allows for MessageBox with custom text scale and overall smaller/larger box size. Useful for popups with a large amount of text that need to shrink a bit to fit on screen.
While other UI windows might also benefit from this kind of thing, I prefer to limit to message box for now as it's a very simple control.