Weapon HUD Animations Mod Support 1, 7/19/2022
Very basic mod support for making it possible through a mod with texture replacements to add unique HUD animation frames for the various different vanilla weapons.

So instead of only being able to replace the entire group of the generic "long-blade" animations that most bladed weapons use, the modder can specify through texture naming if they want to replace the animations for say the Claymore and Katana so they look different from one another on the HUD. Or even adding unique HUD animations for the vanilla weapon artifacts and enchanted magic weapons, etc, 7/19/2022.