Revert strengthens/weakens armor modifier change
I'm reverting after reviewing this change. In short:
1. Strengthens Armor enchant modifies `CalculateArmorToHit` by -5
2. Daedric plate also modifies `CalculateArmorToHit` by -5
At this point in formula, the Strengthens Armor enchantment provides the same modifier as Daedric plate. If this value is also multiplied by 5 here, it will be far too strong.
Yes, this is incongruous with values displayed on paper doll, but paper doll matches classic and internal modifier is equal to Daedric plate, which should also be correct behaviour in formula. This modifier is not as weak as it first seems.
I found one of my conversations about this in below topic a bit closer to implementation time.
https://forums.dfworkshop.net/viewtopic.php?p=43475#p43475
@KABoissonneault I hope that covers it off. It's just another one of those confusing classic setups.