Update PlayerMouseLook.cs
This submission contains a refactoring of the camera smoothing code located in PlayerMouseLook.cs. The purpose of this refactoring is two-fold:

1) To increase behavioral parity between "smoothing" and "no smoothing" modes. Moving the mouse an inch in a certain direction should give identical yaw/pitch deltas (ultimately) for both modes, it's just that one mode is progressive and the other is immediate.
2) To simplify the existing implementation by decreasing lines of code and variables involved while providing the same level of functionality. This wasn't really a goal from the outset but more a result of implementing the first point. Both "smoothing" and "no smoothing" modes are now handled with a single block of code instead of branching into two separate blocks of code.

Note: To increase or decrease the camera smoothing effect you may edit the following value (located at PlayerMouseLook.cs : line 196 at time of original submission):

float smoothing = enableSmoothing ? 0.5f : 0.0f;

For example, a value of 0.3f will give a very subtle smoothing effect and a value of 0.8f will give a very pronounced smoothing effect (at the cost of less responsiveness/higher latency). I have chosen a value of 0.5f because I feel it is suitable for production. It is possible that the "Mouse Smoothing" checkbox could be replaced with a slider ranged 0.0 to 0.9, in future releases.

Note: This submission also contains a small but necessary change in SerializablePlayer.cs (included in this branch).