Remove stubbed BOK string table and related setup
Book files contain custom information such as price and author in addition to actual book text. The challenge is to store this data in a way that is easy to access, easy to view/edit, and easy to replace.
Now using a custom text file format for localized book instead of string table.
Remove stubbed string table setup as never used and no longer required.