Initial support for image markup in localized books
Using a TextLabel override called ImageLabel to draw an image in place of a label inside book reader. This allows image to be treated like a label in terms of size and layout.
Does not seek mods yet. Images must be in StreamingAssets/Text/Books/BookImages.
Image not currently scissored to book reader interior. Will add this soon.