Prevent buttonless message boxes from sticking
Currently, if you create a message box with clickAnywhereToClose set to false (the default), there's no way to get rid of the message box if it doesn't have a button that closes it. Unless there's a reason why this behavior is needed, here's a very small change that will allow the player to close buttonless message boxes.