strip headers using Linq Where
Also, return the lines as a List instead of an Array.

Rationale: Arrays must allocate exactly the number of elements, so
IEnumerable.ToArray() overallocates as it aggregates elements, then as a
last step does an extra copy to return the exact number of elements.

Since the result is short lived, extra allocated elements is not a
concern.

Source:
https://stackoverflow.com/questions/1105990/is-it-better-to-call-tolist-or-toarray-in-linq-queries

Another solution would have been to return an IEnumerable, but it could
generate parsing exceptions from unexpected places.