Improve handling of bad book ID 10000
ID 10000 "Ark'ay The God" is a duplicate of book ID 5.
Removed mapping from books.txt so this book is never added to random pool again. It will still be present in older saves however, so additional handling needed.
When getting book filename, map back to ID 5 so that same file is used in both cases.
When deserializing items, map back to ID 5 on load.
When loading book replacements, explicitly block reserved ID 10000 as this is no longer in books.txt.
Updated comments to reflect that IDs 0-255 and 10000 are reserved by DFU.