Don't disable ClassicUpdate for non-pausing window
Respect the PauseWhileOpen flag on top window and don't freeze ClassicUpdate (where enemy AI runs) if game is otherwise not paused.
If a pausing window is lower in stack, game will be paused and this check will not be reached.
All windows in unmodded DFU will pause game. This test will only be reached for mods where a non-pausing window is just above HUD.